Disable auto backup pause on crossplay enabled server
Do to a recent observation in #1050, the behavior of
AUTO_BACKUP_PAUSE_WITH_NO_PLAYERS should be additionally negated when
cross play is enabled, as the Steam Query port also doesn't return
player counts properly in this situation.

This can be later removed when there is a more concrete method
implemented for querying player counts.