Unify Dockerfile targets (odin/valheim), update release workflow, and modernize dependency stack (#1184)
* Fix deps

* Clippy fix

* Prettier format

* upgrade warp

* Updated Huginn and mod downloader

* [odin] update BepInEx vars to Doorstop 4.x (#1198)

* [odin] remove unused DYLD_* BepInEx vars

These are only used on MacOS and are superfluous here.

* [odin] update BepInEx vars to Doorstop 4.x

BepInExPack_Valheim versions 5.4.2330 and above use Doorstop 4.x, which
uses slightly different environment variables than previous versions.

We update the required BepInEx vars for Doorstop 4.x, while keeping the
older vars present to act as a compatibility layer so that Odin
continues to work with older versions of BepInEx should the user decide
not to upgrade.

* Update Rust crate reqwest to v0.12.23 (#1207)

* Upgrades to bepinex for version tracking and backwards compatability

* log changes

* Fix ci

* Fix up somereadme to trigger ci

---------

Co-authored-by: Nicolas Busseneau <nicolas@busseneau.fr>
Co-authored-by: kodiakhq[bot] <49736102+kodiakhq[bot]@users.noreply.github.com>